mood(updated)/requirements.txt
azure-core==1.32.0
azure-storage-queue==12.12.0
cryptography==44.0.0
Flask==2.0.1
Flask-Cors==5.0.0
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyodbc==4.0.30
pyparsing==3.2.0
python-dotenv==0.19.2
requests==2.32.3
SQLAlchemy==1.4.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==2.0.1
zipp==3.21.0
